+++ packages/calls/src/DidCommCallsService.ts
@@ -7,9 +7,15 @@ import { CallRejectMessage } from './messages/CallRejectMessage'
 
 @scoped(Lifecycle.ContainerScoped)
 export class DidCommCallsService {
-  public createOffer(options: { callType: DidCommCallType; parameters: Record<string, unknown> }) {
-    const { callType, parameters } = options
-    return new CallOfferMessage({ callType, parameters })
+  public createOffer(options: {
+    callType: DidCommCallType
+    offerExpirationTime?: Date
+    offerStartTime?: Date
+    description: string
+    parameters: Record<string, unknown>
+  }) {
+    const { callType, offerExpirationTime, offerStartTime, description, parameters } = options
+    return new CallOfferMessage({ callType, offerExpirationTime, offerStartTime, description, parameters })
   }
 
   public createAccept(options: { threadId?: string; parameters: Record<string, unknown> }) {