+++ packages/calls/src/DidcommCallsApi.ts
@@ -25,13 +25,22 @@ export class DidCommCallsApi {
   public async offer(options: {
     connectionId: string
     callType: DidCommCallType
+    offerExpirationTime?: Date
+    offerStartTime?: Date
+    description: string
     parameters: Record<string, unknown>
   }) {
-    const { connectionId, callType, parameters } = options
+    const { connectionId, callType, offerExpirationTime, offerStartTime, description, parameters } = options
     const connection = await this.connectionService.getById(this.agentContext, connectionId)
     connection.assertReady()
 
-    const message = this.didcommCallsService.createOffer({ callType, parameters })
+    const message = this.didcommCallsService.createOffer({
+      callType,
+      offerExpirationTime,
+      offerStartTime,
+      description,
+      parameters,
+    })
 
     const outbound = new OutboundMessageContext(message, {
       agentContext: this.agentContext,