+++ packages/calls/src/messages/CallOfferMessage.ts
@@ -1,5 +1,7 @@
 import { AgentMessage, IsValidMessageType, parseMessageType } from '@credo-ts/core'
-import { IsString } from 'class-validator'
+import { DateTransformer } from '@credo-ts/core/build/utils/transformers'
+import { Expose } from 'class-transformer'
+import { IsDate, IsString } from 'class-validator'
 
 export type DidCommCallType = 'audio' | 'video' | 'service'
 
@@ -7,7 +9,8 @@ interface CallOfferMessageOptions {
   id?: string
   callType: DidCommCallType
   offerExpirationTime?: Date
-  callStartTime?: Date
+  offerStartTime?: Date
+  description?: string
   parameters: Record<string, unknown>
 }
 
@@ -18,18 +21,28 @@ export class CallOfferMessage extends AgentMessage {
     if (options) {
       this.id = options.id ?? this.generateId()
       this.callType = options.callType
+      this.description = options.description
       this.offerExpirationTime = options.offerExpirationTime
-      this.callStartTime = options.callStartTime
+      this.offerStartTime = options.offerStartTime
       this.parameters = options.parameters
     }
   }
 
   @IsString()
+  @Expose({ name: 'call_type' })
   public callType!: string
 
+  @IsDate()
+  @Expose({ name: 'offer_expiration_time' })
+  @DateTransformer()
   public offerExpirationTime?: Date | null
 
-  public callStartTime?: Date | null
+  @IsDate()
+  @DateTransformer()
+  @Expose({ name: 'offer_start_time' })
+  public offerStartTime?: Date | null
+
+  public description?: string
 
   public parameters!: Record<string, unknown>